+++ src/soeprotocol/mod.rs
@@ -1,0 +1,14 @@
+pub mod protocol;
+pub mod ack_packet;
+pub mod data_packet;
+pub mod disconnect_packet;
+pub mod fatal_error_packet;
+pub mod multi_packets;
+pub mod ping_packet;
+pub mod session_reply_packet;
+pub mod session_request_packet;
+pub mod soeprotocol_functions;
+pub mod soeprotocol_packets_structs;
+pub mod net_status_reply_packet;
+pub mod net_status_request_packet;
+pub mod unknown_packet;
