+++ src/soeprotocol/ping_packet.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+use wasm_bindgen::prelude::*;
+
+#[wasm_bindgen]
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub struct PingPacket {}
+
+
