+++ src/soeprotocol/soeprotocol_functions.rs
@@ -1,6 +1,7 @@
-use super::soeprotocol_packets_structs::*;
 use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
 use std::io::Cursor;
+
+use super::data_packet::DataPacket;
 
 pub enum PacketsMinSize {
     SessionRequest = 14,
@@ -96,7 +97,9 @@ pub fn write_packet_data(wtr: &mut Vec<u8>, data_packet: &mut DataPacket) {
 
 #[cfg(test)]
 mod tests {
-    use crate::soeprotocol::SoeOpcode;
+    use crate::soeprotocol::{data_packet::DataPacket, protocol::SoeOpcode};
+
+
 
     #[test]
     fn write_packet_data_test() {
@@ -105,7 +108,7 @@ mod tests {
         ]
         .to_vec();
         let mut wtr = vec![];
-        let mut data_packet = super::DataPacket::new(data_to_pack, 0, SoeOpcode::Data as u16);
+        let mut data_packet = DataPacket::new(data_to_pack, 0, SoeOpcode::Data as u16);
         super::write_packet_data(&mut wtr, &mut data_packet);
         assert_eq!(
             wtr,
