+++ src/soeprotocol/unknown_packet.rs
@@ -1,0 +1,5 @@
+use serde::{Deserialize, Serialize};
+use wasm_bindgen::prelude::*;
+#[wasm_bindgen]
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub struct UnknownPacket {}
